refactor: 메세지 리스트 컴포넌트 분리
