+++ src/components/MessageList/MessageBox/styles.ts
@@ -1,0 +1,35 @@
+import { css } from '@styled-system/css';
+import { hstack, vstack, wrap } from '@styled-system/patterns';
+
+const msgBox = hstack({
+  p: '1.4rem 2rem',
+  rounded: '0.8rem',
+  border: '1px solid',
+  borderColor: 'gray.30',
+  justifyContent: 'flex-start',
+  fontSize: 'body2',
+  color: 'text.100',
+  gap: '4rem',
+  bg: 'bg.100',
+  w: '100%',
+});
+
+const infoWrapper = vstack({ gap: '1.3rem', alignItems: 'flex-start' });
+
+const info = wrap({
+  gap: '1rem',
+  lineHeight: '1.2rem',
+  '& span': {
+    fontSize: 'caption1',
+  },
+});
+
+const comment = css({
+  color: 'text.200',
+});
+
+const blueText = css({
+  color: 'main.blue',
+});
+
+export { msgBox, info, comment, infoWrapper, blueText };
