+++ src/components/MessageList/styles.ts
@@ -1,0 +1,19 @@
+import { css } from '@styled-system/css';
+import { vstack } from '@styled-system/patterns';
+
+const msgsWrapper = vstack({
+  bg: 'bg.200',
+  alignItems: 'flex-start',
+  gap: '1.5rem',
+  p: '1.3rem',
+  w: '100%',
+});
+
+const title = css({
+  color: 'text.100',
+  fontSize: 'subtitle1',
+  fontWeight: 600,
+  letterSpacing: '-0.02rem',
+});
+
+export { msgsWrapper, title };
